backend/pyproject.toml
[project]
name = "ssepoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi~=0.128.5",
    "uvicorn[standard]~=0.35.0",
    "openai~=1.84.0",
    "google-genai~=1.62.0",
    "python-dotenv~=1.0.1",
    "google-adk>=1.25.0",
]

[dependency-groups]
dev = [
    "ruff~=0.12.7",
]

[tool.ruff]
